api/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.125.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lightgbm==4.6.0
MarkupSafe==3.0.3
narwhals==2.14.0
numpy==2.3.5
packaging==25.0
pandas==2.2.2
pillow==12.0.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.6.1
scipy==1.16.3
six==1.17.0
smmap==5.0.2
starlette==0.50.0
streamlit==1.52.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xgboost==3.1.2
